Next attempt to fix CI
